chore: test reset settings text at the end of settings
