+++ src/components/settings/AppearanceSettings.tsx
@@ -25,7 +25,7 @@ export const AppearanceSettings: FC = () => {
   }, [settings.theme]);
 
   return (
-    <fieldset className="mb-3">
+    <fieldset>
       <legend id="appearance" className="mb-1 mt-2 font-semibold">
         Appearance
       </legend>
