+++ src/routes/Settings.tsx
@@ -1,19 +1,31 @@
-import type { FC } from 'react';
+import { type FC, useContext } from 'react';
 import { Header } from '../components/Header';
 import { AppearanceSettings } from '../components/settings/AppearanceSettings';
 import { NotificationSettings } from '../components/settings/NotificationSettings';
 import { SettingsFooter } from '../components/settings/SettingsFooter';
 import { SystemSettings } from '../components/settings/SystemSettings';
+import { AppContext } from '../context/App';
 
 export const SettingsRoute: FC = () => {
+  const { resetSettings } = useContext(AppContext);
   return (
     <div className="flex h-screen flex-col" data-testid="settings">
       <Header fetchOnBack={true}>Settings</Header>
 
-      <div className="flex-grow overflow-x-auto px-8">
+      <div className="flex flex-col flex-grow overflow-x-auto px-8 gap-3">
         <AppearanceSettings />
         <NotificationSettings />
         <SystemSettings />
+        <button
+          type="button"
+          onClick={() => {
+            confirm('Are you sure you want to reset all settings?') &&
+              resetSettings();
+          }}
+          className="text-sm hover:underline mb-4 hover:cursor-pointer"
+        >
+          Restore settings to their defaults
+        </button>
       </div>
 
       <SettingsFooter />
